+++ modules/util/index.ts
@@ -32,6 +32,7 @@ export { utilGetSetValue } from './get_set_value';
 export { utilHashcode } from './util';
 export { utilHighlightEntities } from './util';
 export { utilKeybinding } from './keybinding';
+export { utilExpandLocaleCode } from './localizer';
 export { utilNoAuto } from './util';
 export { utilObjectOmit, utilCheckTagDictionary, stringifyProperties } from './object';
 export { utilCompareIDs } from './util';